Wireshark - Analyze Your First Packet
Wireshark - Analyze Your First Packet
Scenario
In this scenario, you’re a security analyst investigating traffic to a website.

You’ll analyze a network packet capture file that contains traffic data related to a user connecting to an internet site. The ability to filter network traffic using packet sniffers to gather relevant information is an essential skill as a security analyst.

You must filter the data in order to:

Identify the source and destination IP addresses involved in this web browsing session.
Examine the protocols that are used when the user makes the connection to the website.
Analyze the data packet to identify the type of information sent and received by the systems that connect to each other when the network data is captured.
An overview of the key property columns listed for each packet:

No : The index number of the packet in this packet capture file.
Time: The timestamp of the packet.
Source: The source IP address.
Destination: The destination IP address.
Protocol: The protocol contained in the packet.
Length: The total length of the packet.
Info: Some infomation about the data in the packet (the payload) as interpreted by Wireshark.
Solutions
Identify the source and destination IP addresses involved in this web browsing session.
On the title bar, type ip.addr == 142.250.1.139 to filter for traffic associated with a specific IP address. Select the first packet that contains TCP on the info field. addr means either the source or the destination IP.